Inspector: adding our RunControl to launch apps in debug mode
Our InspectorRunControl is modeled after DebuggerRunControl that stores
the startup parameters of the debugger and then fires it when it's started.

Now we need a clean way to make this new class not interfere with existing
RunControls. You may have 2, one for the original run and another for the
Inspector attachment.

More problems:
 * use existing DebuggerRunControls to apply Inspector over code being
   debugged
 * dispose existing (desktop)RunControls and start (maybe pid-attaching)
   to them with this new RunControl - or just convert them in an unknown (to me)
   way