Inspector: add InspectionTarget and migrate every target description to that
This allows for per-runconf behavior, like enabling or disabling the
start buttons based on the framework and on the current InspectionTarget.

An inspection generates on the DashboawdWindow, then newInspection gets
called with the InspectionTarget and the FrameworkFactory, then the
Framework is created (if possible) and the inspection is started (if
possbile).

The NokiaQtFramework just asks the owned SharedDebugger to start the
InspectionTarget. The SharedDebugger handles the different types of
InspectionTargets (4) and creates RunControls for those and submits
those to the ProjectExplorer which runs everything.

Easy, no? ;-)